requirements.txt
Flask
Flask-APScheduler
Flask-Migrate
Flask-Login
Flask_Reuploaded
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
PyMySQL
psutil
Flask-Mail
sqlparse
Pillow
python-dotenv
validators
redis
reportlab
numpy
